Moved filemanager method tests to test_filemanager.py
